Add the TopNAggregation schema registration
